Add DiscardMapZeros helper transformer
In some situations users want to compare two maps where missing entries
are equal to those that have the zero value.
For example, the following maps would be considered equal:
	x := map[string]int{"foo": 12345, "zero":0}
	y := map[string]int{"foo": 12345}

To help with this, we add DiscardMapZeros to cmpopts that transforms maps
by stripping entries with zero values.

This is not implemented in terms of a FilterPath+FilterValues+Ignore where the
Ignore is filtered to only activate if both of values are either the zero value
or missing. The reason is because the FilterValues will not even be evaluated
for invalid reflect.Value. In other words, there is no way to represent the
concept of a "missing" value in the Filter API.